+++ ShapeClassifier/Rectangle.cs
@@ -5,6 +5,6 @@ using System.Text;
 
 class Rectangle: Shape
 {
-    public double Width { get; set; }
-    public double Height { get; set; }
+    public int Width { get; set; }
+    public int Height { get; set; }
 }
